+++ src/Conceptor.ts
@@ -2,12 +2,12 @@
  * @responsibility Create a webview panel
  * @responsibility List project files according to the include/ignore patterns configured
  * @responsibility Request CRC Cards corresponding to the given set of files
- * @responsibility Request Conception Graph rendering and attach it to the created panel
+ * @responsibility Request Design Graph rendering and attach it to the created panel
  **/
 
 import * as vscode from 'vscode';
 import { CRCCardsGenerator } from './CRCCardsGenerator';
-import { ConceptionGraphGenerator } from './view/ConceptionGraphGenerator';
+import { DesignGraphGenerator } from './view/DesignGraphGenerator';
 import { CRCCard } from './types/model';
 import { ConfigurationManager } from './view/ConfigurationManager';
 
@@ -25,7 +25,7 @@ export class Conceptor {
     );
 
     // Register lifecycle listeners
-    vscode.workspace.onDidSaveTextDocument(this.buildConceptionGraph);
+    vscode.workspace.onDidSaveTextDocument(this.buildDesignGraph);
   }
 
   private static findFiles = async () =>
@@ -34,19 +34,19 @@ export class Conceptor {
       ConfigurationManager.getIgnoreFilePatterns(),
     );
 
-  private renderConceptionGraph = async (crcCards: CRCCard[]) => {
-    this.panel.webview.html = await ConceptionGraphGenerator.withConceptionGraph(
+  private renderDesignGraph = async (crcCards: CRCCard[]) => {
+    this.panel.webview.html = await DesignGraphGenerator.withDesignGraph(
       crcCards,
     )(this.panel, this.context);
   };
 
-  public buildConceptionGraph = async () => {
+  public buildDesignGraph = async () => {
     const fileUris = await Conceptor.findFiles();
 
     const crcCards = (await CRCCardsGenerator.generateCRCCards(
       fileUris,
     )) as CRCCard[];
 
-    this.renderConceptionGraph(crcCards);
+    this.renderDesignGraph(crcCards);
   };
 }