+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to the "conceptor" extension will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2020-04-27
+
+### 🚑 Fixed
+
+- Fix graph rendering (copy rendering templates during compilation)
+- Rename "Conception" into "Design" for more accuracy
+
 ## [0.2.0] - 2020-04-22
 
 ### 🎉 Added
 
-- Live update the conception graph on each file save
+- Live update the design graph on each file save
 
 ### 🦋 Changed
 
@@ -20,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🎉 Added
 
 - Provide include/ignore file patterns configurability
-- Provide conception graph layout configurability
+- Provide design graph layout configurability
 
 ### 🦋 Changed